requirements.txt
flair==0.7
discord.py==1.6.0
pandas==1.2.1
